pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nullcode"
version = "1.0.0"
description = "AI-powered CLI vulnerability scanner for ethical hackers"
authors = [{name = "Null Security Team"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["security", "vulnerability", "scanner", "cli", "ai", "ethical-hacking"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
]

dependencies = [
    "torch>=2.2.0",
    "transformers>=4.36.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "tqdm>=4.66.0",
    "semgrep>=1.50.0",
    "gitpython>=3.1.0",
    "pyyaml>=6.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "black==23.12.1",
    "mypy==1.8.0",
]

[project.scripts]
nullcode = "nullcode.__main__:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["nullcode*"]

[tool.black]
line-length = 100
target-version = ["py310"]
